requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.5
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-nose==1.4.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==8.1.2
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
nose==1.3.7
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.4.0
Pillow==8.2.0